Rename to v2 and use apiextensions
