+++ service/cloudconfigv2/cloud_config.go
@@ -1,13 +1,14 @@
-package cloudconfigv1
+package cloudconfigv2
 
 import (
-	"github.com/giantswarm/aws-operator/service/keyv1"
-	"github.com/giantswarm/awstpr"
+	"github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/certificatetpr"
 	cloudconfig "github.com/giantswarm/k8scloudconfig"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/randomkeytpr"
+
+	"github.com/giantswarm/aws-operator/service/keyv2"
 )
 
 const (
@@ -53,18 +54,18 @@ func New(config Config) (*CloudConfig, error) {
 
 // NewMasterTemplate generates a new worker cloud config template and returns it
 // as a base64 encoded string.
-func (c *CloudConfig) NewMasterTemplate(customObject awstpr.CustomObject, certs certificatetpr.CompactTLSAssets, keys randomkeytpr.CompactRandomKeyAssets) (string, error) {
+func (c *CloudConfig) NewMasterTemplate(customObject v1alpha1.AWSConfig, certs certificatetpr.CompactTLSAssets, keys randomkeytpr.CompactRandomKeyAssets) (string, error) {
 	var err error
 
 	// Default the version if it is not configured or we are using Cloud Formation.
 	// TODO Remove once Cloud Formation migration is complete.
-	if !keyv1.HasClusterVersion(customObject) || keyv1.UseCloudFormation(customObject) {
+	if !keyv2.HasClusterVersion(customObject) || keyv2.UseCloudFormation(customObject) {
 		customObject.Spec.Cluster.Version = string(cloudconfig.V_0_1_0)
 	}
 
 	var template string
 
-	switch keyv1.ClusterVersion(customObject) {
+	switch keyv2.ClusterVersion(customObject) {
 	case string(cloudconfig.V_0_1_0):
 		template, err = v_0_1_0MasterTemplate(customObject, certs, keys)
 		if err != nil {
@@ -80,12 +81,12 @@ func (c *CloudConfig) NewMasterTemplate(customObject awstpr.CustomObject, certs 
 
 // NewWorkerTemplate generates a new worker cloud config template and returns it
 // as a base64 encoded string.
-func (c *CloudConfig) NewWorkerTemplate(customObject awstpr.CustomObject, certs certificatetpr.CompactTLSAssets) (string, error) {
+func (c *CloudConfig) NewWorkerTemplate(customObject v1alpha1.AWSConfig, certs certificatetpr.CompactTLSAssets) (string, error) {
 	var err error
 
 	// Default the version if it is not configured.
 	// TODO Remove once Cloud Formation migration is complete.
-	if !keyv1.HasClusterVersion(customObject) {
+	if !keyv2.HasClusterVersion(customObject) {
 		customObject.Spec.Cluster.Version = string(cloudconfig.V_0_1_0)
 	}
