+++ service/cloudconfigv2/v_0_1_0_worker_template.go
@@ -1,13 +1,13 @@
-package cloudconfigv1
+package cloudconfigv2
 
 import (
-	"github.com/giantswarm/awstpr"
+	"github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/certificatetpr"
-	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v_0_1_0"
+	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v_2_0_0"
 	"github.com/giantswarm/microerror"
 )
 
-func v_0_1_0WorkerTemplate(customObject awstpr.CustomObject, certs certificatetpr.CompactTLSAssets) (string, error) {
+func v_0_1_0WorkerTemplate(customObject v1alpha1.AWSConfig, certs certificatetpr.CompactTLSAssets) (string, error) {
 	var err error
 
 	var params k8scloudconfig.Params
@@ -37,7 +37,7 @@ func v_0_1_0WorkerTemplate(customObject awstpr.CustomObject, certs certificatetp
 
 type v_0_1_0WorkerExtension struct {
 	certs        certificatetpr.CompactTLSAssets
-	customObject awstpr.CustomObject
+	customObject v1alpha1.AWSConfig
 }
 
 func (e *v_0_1_0WorkerExtension) Files() ([]k8scloudconfig.FileAsset, error) {
